coa_assistant/requirements.txt
streamlit
pandas
sentence-transformers
faiss-cpu
llama-cpp-python
